CRUD Boletim

+++ 2 - Domain/Entities/Boletim.cs
@@ -9,7 +9,7 @@ namespace CursoIdiomas.Domain.Entities
     public class Boletim : Entity
     {
         public Boletim() { }
-        public Boletim(float nota, long idAluno) {
+        public Boletim(double nota, long idAluno) {
             Nota = nota;
             DataPublicacao = System.DateTime.Now;
             IdAluno = idAluno;
@@ -20,7 +20,7 @@ namespace CursoIdiomas.Domain.Entities
 
         public DateTime? DataPublicacao { get; private set; }
         public DateTime? UltimaAtualizacao { get; private set; }
-        public float Nota { get; private set; }
+        public double Nota { get; private set; }
 
         public long IdAluno { get; private set; }
         public Alunos Aluno { get; private set; }
